requirements.txt
Pillow==9.3.0
ujson==5.4.0
